requirements.txt
streamlit==1.37.1
pymongo==4.7.2
python-dotenv==1.0.1
bcrypt==4.1.3
pandas==2.2.2
matplotlib==3.9.0
